InputDispatcher: Refactor CommandQueue to use lambdas
InputDispatcher uses a CommandQueue to post messages that should be run
later on the dispatcher thread, similar to a Handler.

Instead of using function pointers and `CommandEntry`s to enqueue messages
to the queue, we use use lambdas that also have a copy of the parameters.

Doing this reduces some redundancies and verbosity, as well as
increasing readability.

Bug: None
Test: atest inputflinger_tests
Change-Id: I330cbaed2532228e7e58b42b6ce4a079328cdd97