Add support for latching unsignaled buffers in Skia RenderEngine.
When we re-added support for latching unsignaled buffers we forgot to
propagate this support to Skia RenderEngine, since we were assuming that
latching unsignaled buffers would no longer be supported.

Because partners are still relying on latching unsignaled buffers being
available for good performance in some use-cases, Skia RenderEngine
needs to wait on incoming buffer fences to prevent tearing.

Bug: 196201621
Bug: 195033578
Test: atest librenderengine_test
Change-Id: I1f3be882e55de20bdbfa3bf92bf0b8f88b8bc3ea