Use multiplied timeout for tests
Currently, the timeout value is hardcoded. But this poses a problem for
setups where the timeout is different from the default value.

One such example is hwasan. It turns out that, likely recently, the hw
multiplier was set to 2 for hwasan builds. This causes the
inputflinger_tests to fail on hwasan platforms.

To fix this, use the same timeout that InputDispatcher uses.

Bug: none
Test: atest inputflinger_tests
Change-Id: Ibf752d90d03765bbf6ed2b6d9c10e8783c9f3564