Remove displayBounds when computing layer bounds
Display bounds will be applied later in CE and RE so there's no need for
them to be used when computing the layer bounds. This fixes a few
issues:

1. Accessibility can now get the real bounds of the window, even if it's
offscreen when magnified

2. Screenrecord rotation flicker is fixed. This was caused by the
multiple calculation using different display bounds and caching the last
value

3. Reducing number of calls to computeBounds. We now only call it once
where previously we'd call it once per display. Each computeBounds
traverses down the entire hierarchy so it can be inefficient.

Test: Existing tests pass. WindowInfo has bounds outside display
Bug: 188792659
Fixes: 140855415
Change-Id: I1f5072adb674c7c4dacb6a392473821968c67767