Use a FakePointerController for InputReaderIntegrationTests
InputReaderIntegrationTests uses the real EventHub when running on a
device, which means the test is affected by the input devices connected
to the test device.

If the test device has a cursor connected to it, the test will try to
obtain a FakePointerController from the FakeInputReaderPolicy. If the
pointer controller is not set, then the null pointer that is return will
cause a NPE. We set the PointerController when setting up the
integration tests.

Here, we also remove the need to have a per-device PointerController in
the tests, because there should only be one PointerController that's
shared by all input devices.

Bug: 195058024
Test: forrest test run with CL
Change-Id: I937c6fead36cf648ca38e8a0d0d662b2188fb241