Add applyPendingTransactions
It's possible for a caller to request mergeWithNextTransaction, but
the main frame had nothing new to draw. If that's the case, the
mergeWithNextTransactions will be stuck and never applied (or applied
much later). Since this could end up blocking Transactions, it's better
to force apply these when we know a frame wasn't going to draw this
vsync.

Adding applyPendingTransactions to allows callers to force apply
these transactions.

Test: Existing tests pass
Bug: 195262673
Change-Id: I3082b54e35dfae2b3f7fe589c6f665f781d8b07b