RefreshRateConfigs: fix getMaxRefreshRateByPolicyLocked
If there are no layers which are requesting a
non seamless switch getMaxRefreshRateByPolicyLocked
should return the max refresh rate from the group of
the default mode (and not the group of the current mode).

This fixes a bug where when the display mode is never
reset to default after we exit an app which has called
setFrameRate with  changeFrameRateStrategy=Always.

This change should have no affect phones, because
there all modes are in the same config group.

Bug: 199270559
Test: atest MatchContentFrameRateTest SetFrameRateTest
Test: atest RefreshRateConfigsTest
Change-Id: I6ff200a85edf026b0d04d4a06a13995832ffdab4