MotionEvent: Report transformed orientation values
Previously, the AXIS_ORIENTATION value reported by MotionEvent was
unaffected by the window transform. This meant that whatever value that
was generated by InputReader was the value reported to the window. This
meant that if the window was rotated or scaled respective to the logical
display and the window received an event, it could be the case that the
0 value no longer corresponds with the "up" direction of the window.

Now that the input pipeline works in the display panel's coordinate
system, we need the orientation value reported by MotionEvent to be
affected by the window transform.

This CL also refactors the shared logic by which transformed axis values
are calculated into a common function.

Bug: 179274888
Test: atest libinput_tests
Test: atest inputflinger_tests
Test: manual with test app and stylus
Change-Id: Ibb6f135de47f7c1cbde3c023931a760dfbe08e45