Clean up KawaseBlurFilter
The clamp() calls aren't necessary with the current design and since
x/y are treated the same any duplicated code around stepX/stepY can
also be deleted.

This makes the shader a little simpler and reduces the number of
intermediate const variables that are passed in twice for x and y.

The reason for doing this is to prepare to use this code as a base
for other implementations of blur.

Bug: 185365391
Test: atest BlurTests
Change-Id: I6c75d89dd10d568651f6a6183d25b170ab8526ad