Remove some restorcon operations
Change cache and code-cache dir restorecons to only work on creation
of these dirs instead of doing it everytime. Currently, we call
the restorecon for the cache and code-cache dirs each time createAppData
is called (i.e. when a user is started or created). This change
removes the restorecon calls to these dirs on user start,
thereby saving ~500ms on the call. Additionally, there are changes to
force a recursive restorecon of directories that are newly created
to save an additional lookup that's done to check if the label has
changeed.

Bug: 191997617
Test: Remove the cache and code-cache dirs for a particular
app and user using adb shell. Reboot the device and switch to
that user. Check that the cache/code-cache dirs are re-created.

Change-Id: Ic7ff0f8181cb06ff9eb4b60ded92d0c8be31bd2b