Move sysprop check for RenderEngineType into SF
As brought up in I455d5d613ccaa210dc748969e025dc86c78080b8, we should
read the sysprop in SF, so that clients (e.g. tests or benchmarks)
that specify using a particular Type actually use that type, regardless
of any sysprop set.

If the sysprop is not set, or is set to an invalid value, just respect
the default that is already set by RenderEngineCreationArgs::Builder.

Bug: 193240340
Test: adb shell setprop debug.renderengine.backend
      verify via logcat that the proper type is used

Change-Id: Iaf425fde6333f54d5b190df66cb7a79e1d491e63