BBQ: Clean up acquire states on BQ disconnect
When the producer disconnects, all buffers in the queue will
be freed. So clean up the bbq acquire state and handle any
pending release callbacks. If we do get a release callback for
a pending buffer for a disconnected queue, we cannot
release the buffer back to the queue. So track these
separately and drop the release callbacks as they come.

Transaction callbacks are still expected to come in the
order they were submitted regardless of buffer queue
state. So we can continue to handle the pending
transactions and transaction complete callbacks. When
the queue is reconnected, the queue will increment the
framenumbers starting from the last queued framenumber.

Bug: 201482894
Bug: 197269223
Test: atest BLASTBufferQueueTest
Change-Id: I36791661e2b6112401d85f9b50495931d82125e4