Reland "Allow changing composition from DISPLAY_DECORATION to DEVICE"
The first time this topic landed, it resulted in b/212402133. We avoid
running into this bug with Ib11d46439db57b90486bad07dd90f2cf0822182a.

Original commit message:

After some discussion, we've decided to always set the ScreenDecorations
to DISPLAY_DECORATION. HWC can decide, based on the format, whether to
truly treat it as DISPLAY_DECORATION, or to change to DEVICE
composition.

Previously we were concerned that SurfaceFlinger couldn't use this
information (if we provided it). Looking to the future, it's possible SF
may use this information to help make a better plan for the next frame.

Bug: 193170859
Test: manual

Change-Id: Ib67f55b11a8c5de8f5763394b00fffee43557ed8