Add integration test that clarifies behaviour of ACTION_POINTER_UP
This test shows what will happen when a pointer goes up. For example, if
a pointer goes up at the same time as another pointer moves, which will
happen first, MOVE or POINTER_UP?

This integration test explores these behaviours. We need to know this
behaviour because later, if we want to suppress a pointer, we need to be
able to cancel pointers without losing information about motions of
other pointers.

Bug: 198472780
Test: atest inputflinger_tests
Change-Id: Ie56cca1324ffdd777eecbd0c6a2fe413d59f241a