gralloc: Move validation of non-blob/gpu_data_buffer down to libui
AIDL gralloc implementations may choose to support this combination.
Existing implementations of HIDL gralloc 2/3/4 do not support it,
and do not reliably reject it due to having been previously protected
by validation at the AHardwareBuffer layer.

Move the check down to the gralloc layer, and defer to the HAL only
in the AIDL case.

Bug: b/195944622
Test: build; real CTS/VTS to land soon after.
Change-Id: I7b0b9de005a9b0f564915f9c71cc8aea3ef0618a