Sync native sensor privacy aidl with frameworks/base
The new calls no longer require user ids since that is managed
internally in the sensor privacy service and calling services will
receive updates when the state is changed or user is changed who has a
different setting. Calling services should treat the state as a global
setting.

Also the new calls add support for hardware and software versions of
controls.

Test: Build
Bug: 191745272
Change-Id: I08c8fa729b478773dc7f1fe7c2a5c22dcd5f0d4c