Upadate sensor service to use new sensor privacy api
The new calls no longer require user ids since that is managed
internally in the sensor privacy service and calling services will
receive updates when the state is changed or user is changed who has a
different setting. Calling services should treat the state as a global
setting.

The new calls also support hardware and software types of controls. This
change treats both the same.

Test: atest CtsSensorRatePermissionTestCases
Bug: 191745272
Change-Id: Ifa905110b0b75fbc48a37b7e49e269621f03129d