SF: Configure input windows with the first input-enabled display
We can have a situation where more than one DisplayDevice can receive
input for a layer stack. In this situation, we don't have a clear way to
distinguish which display we should use for input. We change the
implemntation so that we use the first input-enabled display in
mDisplays to configure the input windows.

In the case where a display is mirrored using the same layer stack, this
would help if the display we actually want to receive input is listed
first.

Bug: 219191788
Test: manual: Developer options > "Simulate secondary display"
Change-Id: Ic9cfff75f0178f80decf6cf1ffd3ba81a766caa6