Enable drawing layers' borders in SurfaceFlinger.
Provide an API on SurfaceComposerClient::Transaction for enabling
a border to be drawn for a specified hierarchy. The caller requests a
border to be drawn at a certain Layer and all its children will be
included when computing the visible region. The border will draw around
the union of all the layers' visible regions, starting from the
requested layer.

Test: go/wm-smoke
Test: LayerBorder_test
Bug: 225977175

Change-Id: I7760b51b68cdf01bb9146ec91a270a9d63927995