SurfaceFlinger: Remove blocking behavior of animation transaction.
Animation transaction is only used in one place, and the blocking
behavior wont be useful since we also use the vsync id. The blocking
behavior has a weird quirk where we only block if the transaction is
in mPendingTransactionQueues and not in mTransactionQueues. It's not
clear if this was even totally intentional, but it's tested in some unit
tests. An upcoming CL to merge mPendingTransactionQueues with
mTransactionQueues will cause some complication to this. The simplest
fix is to just remove the blocking behavior like this, which should also
carry some performance benefits. It would be nice to just totally remove
animation transactions but the flag is still used to enable some
specific stats, which while they are themselves probably not widely
used, I ran in to some issue the last time I tried to remove them.

Bug: 220929926
Test: Existing tests pass
Change-Id: Ib8ce5bff86a53bcac9999b30138c9721a83b6e9b