Reset intercept key wake time when receving up event
The policy need to handle the key shortcuts via policy callbacks,
and it may have a pending timeout in order to detect the combination
keys. If the pending key event has already known an up event has been
queued, we could reset the wake time to speed up the processing for it
may be handled in policy or not.

Bug: 220669913
Test: atest inputflinger_tests
Change-Id: Ic002b770632e15506fd2f581fc8716b26ea3626c