Include SDK sandbox while calculating app and user size
SDK sandbox storage has been separated and need to be specially included
while getting app size and user size.

While calculating app size manually, each of the subdirectories in the
sdk sandbox storage (shared, directories for each sdk etc.) are added up
separately to the total. This is because each of these subdirectories
would contain a cache storage that needs to be added to the cache
calculation.

For quota calculation, the uid of the sandbox storage is used. Cache
data is already added by default to the cache size of the app since the
gids are the same.

Bug: 215506889
Test: atest SdkSandboxStorageHostTest#testSdkDataIsAttributedToApp
Ignore-AOSP-First: Will cherry pick to AOSP once dependent CLs are also
cherry-picked.

Change-Id: I94bc30f7e47e06b3b60d28f16a647a2a0daaae93