Update IfExepensiveOutputDataspaceIsUsed test
Prior to I5c77769cf9fd2f07806c97022fa4bb940bfcc1ac, using a P3 output
space meant we would boost the GPU frequency. This was a shortcut for
having a different dataspace, since it would require more GPU work. The
more thorough check means that we boost only if the inputs don't match
the output. Update the test to include a single layer, which defaults to
a different dataspace, making the test pass again.

Bug: 223651281
Test: libcompositionengine_test
Change-Id: Iad084a8388e6203b83135ebfb632fda1b3cdb448