Add dynamic debug logs to VelocityTracker
This will allow us to debug any issues related to the switch to the
'impulse' velocitytracker algorithm.

In addition, print the lsq2 velocity values for the case when 'impulse'
is selected, and when debugging is enabled. This will make it simpler
for us to follow up on dogfooder's feedback.

Bug: 134179997
Test: atest libinput_tests:VelocityTrackerTest#ThreePointsLinearVelocityTest
Test: (enable debug logs and make sure the impulse and ...
Test: lsq2 velocities match for the test case above)
Change-Id: I3c523b67a0cebe8adaddba8f86d335a2f8f4d8f8
(cherry picked from commit 276467bdf6dc1ce1f3fa9623127e4d298bbc0f2d)