Fix transparent region tests
Follow up to I1378c96f4d5af57fb7d9835182427017064e89e5. Instead of
modifying kTransparentRegionHint, which is already used by other tests,
add kTransparentRegionHintTwo specifically for this test. The original
doesn't work because we want to verify that this test rotates the
result, so we need a rectangle that is not a square.

Fixes: 225841766
Test: libcompositionengine_test
Change-Id: I132cbb15520b76396851673ec33dc9b7cc183db0