Use CacheId when buffer is null.
It's likely the buffer is null even if eBufferChanged is set because we
already cached the buffer. This can happen when the transaction is sent
over IPC and we automatically cache the buffer.

This change uses the cacheId if the buffer is null

Test: Hard to repro
Bug: 225815507
Change-Id: Ic95bfc179cc86e494943e9bd6e15cb9afd69f1df
(cherry picked from commit dcba9e07976c20bb98a3e92aca149eacc6e59d2d)
Merged-In: Ic95bfc179cc86e494943e9bd6e15cb9afd69f1df