[SF] Add a new method for latch unsignaled transactions
When Unsignaled transaction is applied, buffer processing may
stall. The display appears frozen, and recording a trace is
hard or not much use. It can also be hard to reproduce these
scenarios.

Add a new method which helps identify unsignaled transactions in
stacktraces. It calls the same method as before, so there is no
change in behavior.

BUG: 221079645
Test: atest libsurfaceflinger_unittest, atest libgui_test, atest VtsHalGraphicsComposer3_TargetTest
Change-Id: I5ad46e7c3f2784091b7d973f16056de9c8a4ac4b