Check for IOCTL feature changed to not set inheritence
Inheritence flags should only be set on directories and not files,
setting inheritence flags on files fails and leads to not using project
ids.
Also refined the tests to make sure that project ids are indeed used.

Bug: b/215154615
Test: atest installd/StorageHostTest
Test: atest installd/installd_service_test.cpp
Ignore-AOSP-First: It will be picked up in AOSP manually, needs to be in
master first to resolve merge conflicts.
Change-Id: Iac38ba6f77c14141d8b06b63f52bf75fe0c54c66

Change-Id: Ia5baf127af5a078592d8b2db4b2f3d8c048a8692