Treat dimming-disabled layers as HDR
Screen rotations currently capture an SDR screenshot, and rotate the
screenshot. If there is an HDR layer on screen, this causes the HDR
layer to be tone-mapped to SDR, and the screeenshot when fed back into
SurfaceFlinger has a tag that prevents the screenshot from being dimmed
so that the HDR layer can remain at the same brightness (modulo a
brightness change). But this means that the screenshot should be treated
as an HDR layer, as there is a noticeable dip and re-ramp up in
brightness.

Bug: 228197661
Test: builds, boots
Change-Id: I1d1b0dcaf230300ca34b84ea407d0817feb2c664