Add a minimum physical threshold for freeform gesture on touchpad
Currently, TouchInputMapper uses the product of the size of the touchpad
and a ratio (currently 0.25) to determine a two finger gesture is a
freeform gesture or a swipe, that when the distance of two pointers is
greater than the calculated value, it is a freeform gesture, otherwise,
it is a swipe gesture. However, on a really small touch pad, such as
Sony Dualshock4's, due to the common size of human fingers, almost all
two finger gestures are decided to be freeform. Adding a minimun
physical value for the freeform gesture, 30 mm, can solve the problem.

Bug: 197146318
Test: manual test
Change-Id: Ia4ff521cb2f47b587fa55a5d753b90d023a648ef