Disable diag in libinput
The diag for integer sanitizer was originally added to help debug some
of the sanitizer errors (incorrectly using size_t instead of ssize_t
did cause a crash, but it took a long time to understand what the issue
is because the stack wasnt working well).
To reduce the memory usage by some vendor processes, disable diag for
integer sanitizer.

Bug: 228251032
Change-Id: I5d394f9674b49104c22510e136eebbe45e5f0018
Test: adb shell showmap \`pidof system_server\`