SF: include number of Skia shaders in perfetto traces
When a new shader is compiled in SurfaceFlinger (tracked by the ATRACE
tag "shader_compile"), SkSLCacheMonitor::store will be called.
(Eventually, we'd like to store the shader in a persistent cache - this
is blocked on b/119203607.) Use this method to keep track of the total
number of shaders used since boot. Print this number in the perfetto
trace.

This can be used to verify whether we have compiled enough shaders to
require cache eviction, and can also be used to determine how much
larger the cache should be. We are currently using
GrContextOptions.fRuntimeProgramCacheSize's default of 256.

Bug: 226662927
Bug: 222355787
Bug: 222109133
Bug: 190487656
Test: perfetto trace
Change-Id: Ifa99a8a2e16c136dc6f19e2a8dda0692aa4358a8