Move common Skia code to SkiaRenderEngine
This CL takes the code in SkiaGLRenderEngine that is not GL-specific and
moves it to SkiaRenderEngine. This is a pure refactoring, no change in
functionality. This is to prepare for adding a SkiaVk backend.

We introduce virtual functions waitFence, flushAndSubmit, and
backendDump to SkiaRenderEngine so that the same GL fence flush logic
can be called from drawLayers in SkiaRenderEngine, and that most of the
dump() logic can also stay in SkiaRenderEngine.

That leaves SkiaGLRenderEngine doing only EGL context initialization and
fence handling.

Test: librenderengine_test
Bug: 236390072
Change-Id: I240d1723d28836a050b6c694bc61011311106937