libbinder_ndk: no min sdk version - avoid weak
'stability.o' is not included in the build from libbinder_ndk.a when
we statically include it into fuzzers because AIBinder_markVintfStability
is a weak symbol in this case. This is generally not a problem because
libbinder_ndk is built as a shared library. However, for host fuzzers,
we need this symbol to avoid the runtime trying to jump to 0x0.

Fixes: 230228062
Test: android.hardware.vibrator-service.example_fuzzer
Change-Id: Ie8452e2610f3046d9aa5be094ac1dfd1aec8be08