Do not transform values from a SOURCE_MOUSE_RELATIVE device
In Pointer Capture mode, we want the API to report events directly
as they are coming from the kernel without any transformations.
We already do this for SORUCE_TOUCHPAD. Here, we update the handling
of SOURCE_MOUSE_RELATIVE so that when Pointer Capture is enabled,
mouse events are not effected by the display or window state.

Bug: 233184154
Test: Apply non-standard scaling to a device using
"adb shell wm size" and run: atest MicrosoftSculpttouchTest
Test: atest libinput_tests
Test: atest inputflinger_tests

Change-Id: Ia6ef5a21312da949ac2704e9bfe4638626d3a894