Classifier: Ping the HAL from monitor() instead of dump()
The input dump should not be affected by the state of the HAL.
Previously, we were pinging the HAL duing a dump, which means the
dumpsys would fail if the HAL is stuck.

Instead, we ping the HAL duing the monitor() call so that a HAL that is
stuck can be more easiliy identified.

Bug: 237347585
Test: manual, flash device
Change-Id: I050bacc5510f10b788232e156ea91d505d841b55