Switch to resample_period instead of bool resample_touch
When resampling code got upstreamed, we had to switch the 'api surface'
from using a regular bool to providing a value for the resampling
period.

Reference:
https://chromium-review.googlesource.com/c/chromium/src/+/3689765/2..8/ui/events/ozone/evdev/touch_filter/neural_stylus_palm_detection_filter_model.h#b102

Update input code accordingly.
To avoid having to include base::TimeDelta inside input code, in this CL
im adding a custom define of the model period to the chrome's palm
rejection code.

We can try to upstream that again in the future.

Bug: 237238491
Test: build only
Change-Id: I3d57058f45b63201dc363fc0c2528e3f90090244