Fix drag and drop access wrong pointer id
If the touched window didn't support split, it won't track the
pointerIds. When the drag and drop started, it would rely on the
pointerIds to check current touch state and store the initial drag
pointer id but it would always be 0.

This CL will let pointerIds could track all down pointers when touched
windows received down or pointer down, so it could access the right
pointer id from current touch state.

Test: atest inputflinger_tests CrossAppDragAndDropTests
Bug: 237233207
Change-Id: Ia5d91814a2aca56095c29b029e35e66cd669bce9