Call Filter from a separate function
The function 'processMotion' has gotten too long. To make it easier to
reason about, separate some code from the center of the function into a
new function.

Unfortunately, this new function will have side-effects, but it's still
easier to understand and reduces the number of local vars.

There's no functional change in this CL.

Bug: 241935838
Test: atest inputflinger_tests
Change-Id: I2157798b70299659791ae6ef85b2394e9130b4b3