Do not read kernel configs on the host
When the code runs on host (so that it can validate the key layout
maps), it currently will attempt to read the kernel configs on the host.
What we actually want is to either read the kernel configs on the
device, or to always assume that the config exists.

In this CL, we simply assume that the config exists. Since we are
calling this tool for all key layouts, it should not affect the
validation.

Test: m out/target/common/obj/ETC/validate_framework_keymaps_intermediates/stamp
Bug: 237835584
Change-Id: I79158a7591fa4d8ff0a33dcae9df2b138ae33485