Delete TouchInputMapper DeviceType::TOUCH_PAD
This doesn't appear to do anything — it is set in a couple of cases
(notably not ones that actually match with any known "touchpad" devices,
which don't have REL_X or REL_Y), but no other code refers to it. Maybe
it's related to "touch navigation" devices, but they already have a
separate TOUCH_NAVIGATION device type.

Bug: none
Test: atest inputflinger_tests
Change-Id: I786a3d29fa893cb982873aea7695d754b2c096c6