Revert behavior of loading the AssociatedDevice
ag/19714571 changed the behavior of loading AssociatedDevices so that it
was only read once for each sysfsPath, even if multiple event hub
devices had the same sysfs path.

However, that caused CTS hardware tests that read battery info to be
flaky. To mitigate this, we revert to the previous behavior of
re-loading the AssociatedDevice for each event hub device for now, until
we can debug the source of this issue.

Bug: 243979881
Test: atest SonyDualshock4BluetoothTest#testBattery --iterations 100
Change-Id: If7ab1076f0528945cf75073986ad0397471533ca