Turn down GLESRenderEngine tests
Some tests are flaky which is causing delays in continuous integration.
GLESRenderEngine has not been the default backend for RenderEngine since
Android 12, and no known OEM currently opts out of Skia's RenderEngine,
which is the replacement. So, now's not a bad time to turn down these
tests.

Also, as part of removing the tests, turn on and fix some color
management tests which had been accidentally disabled.

Bug: 245359125
Bug: 199918329
Test: librenderengine_test
Change-Id: I1ec93eaad63daa2268f92403bf71bcb65861072b