Improve logs for windows with NO_INPUT_CHANNEL
When touched, these windows today generate a misleading log. The log
states that the window is not responsive. In reality, these windows
simply don't have an input channel.

Touches should continue to be dropped for such windows, but the logs
should be improved.

In this CL, the logic is flattened. We explicitly check for whether the
window has NO_INPUT_CHANNEL flag and generate the appropriate log.

Bug: 249163063
Test: atest inputflinger_tests
Change-Id: I1858c5f10d9a9c6a6916691a72d9e73486480311