Delete struct PointerData to fix ODR violation
This data structure is defined with the same name in
PreferStylusOverTouch_test and in UnwantedInteractionBlocker_test. This
creates a "one-definition-rule" violation.
This violation is not getting caught by the compiler today.

As a result, this causes the read to go out of bounds when the wrong
data structure is used. The issue can be reproduced by running the tests
on host.

To fix it, simply remove this struct and fallback to the Point struct
from libui.

Bug: 217165277 249591924 251318977
Test: atest --host --no-bazel-mode inputflinger_tests
Change-Id: I79e0647df63bbf69dd3074f1bd2931b298f4cd48