Split up TouchInputMapper::preparePointerGestures
This monster 700-line method was quite difficult to understand. It still
is, but hopefully splitting it up into a few more methods helps.

Also, activeTouchId was basically just being used as a shorter name for
mPointerGesture.activeTouchId (with both being assigned to at once),
leaving space for a bug to come in if an assignment was made to one but
not the other. Making it a const reference removes that possibility, and
makes it a bit clearer what's going on.

This CL should not result in any behaviour changes.

Bug: none
Test: atest inputflinger_tests
Test: compare touchpad gesture behaviour before and after
Change-Id: Ia85bf5dd80d87bd8fb83332f50dc25ad8769c32e