Update IsZygote to return true for webview_zygote
This change adds "webview_zygote" to the list of processes that the
IsZygote() function will return true for. This is intended to reduce the
time spent during the "DUMP TRACES" section of dumpstate, which has been
seen to always time out on the WebViewZygoteInit process.

Over two trials of health/bugreport/capturebugreport on bluejay, this
was seen to have reduced the duration of "DUMP TRACES" by 7.474s
(-35.95%) and the duration of the entire bug report by 6.939s (-8.32%).

Test: Build and flash; verify that health/bugreport/capturebugreport
      demonstrates that
      1) the WebViewZygoteInit process is no longer captured by
	 libdebuggerd_client
      2) the duration of the dump traces section is shorter
Bug: 253670403
Change-Id: Ib1ea2f16c7df97181e98d0723d5ed25415f2d48f