SF: Fix onSurfaceFrameCreated for layers without buffers
Previously we did not create surfaceframes for layers
without buffers. This behavior regressed with flattening
of the layer types and causes some jank classifications
to be eviceted from mPendingJankClassifications list
which has a max value of 25.

Fix this by checking if the layer has a buffer and also
while we are here, increase the max to 50 in anticipation
of multiple displays and multiple layers.

Test: presubmit, logcat for no logspam
Bug: 253555650
Change-Id: I0be3d7706305e059074b9efc14428316bdb41e6d