Use getrandom for device input event ID generation
On device, let's go back to using 'getrandom' for the generation of
random input event ids. This should be faster than reading from
/dev/urandom.

This is a partial revert of the previous patch. After this patch, we
will only use '/dev/urandom' on host.

We may be able to switch to 'getrandom' for both if we ever switch to
musl for host. We can revisit it then.

Bug: 254215895
Test: m &&  atest --host libinput_tests
Change-Id: Idc6241facbd93f3f71ae90567db831d96a5fc98b