Remove TouchState::split variable
Instead of storing this state and mutating it in various places, let's
replace the variable with some logic about whether a touch should be
split.

Ideally, we wouldn't even have custom logic for split vs non-split, but
this is a starting point before we can refactor the code further.

The short-term goal is to simplify the findTouchedWindowTargetsLocked
function.

Bug: 211379801
Test: atest inputflinger_tests
Change-Id: I799c2e11a72052a2867710593b18e05011ca44b6