Do not assume we are in POINTER mode when canceling pointer gestures
The change in ag/19758680 made it so that we now reset TouchInputMapper
after the viewport is disabled.

When resetting, we cancel any ongoing gestures. Since the mapper is
re-configured before the gesture is canceled, this means that as the
viewport is getting disabled, the mapper (which was previously in
POINTER mode) is set to DISABLED mode. When the pointer gesture is
canceled afterwards, it is happening while the mapper is already
DISABLED.

In this CL, we allow pointer gestures to be aborted even when the mapper
is not in POINTER mode.

Bug: 257071757
Test: atest inputflinger_tests
Change-Id: I5c80a5c1c411d16f70ed4f7cce6dd97ed91e124f