Remove TouchState::displayId
TouchStates are already per-display, so we don't need to store the
display id inside them.

To check if touch state is "valid", we can just check if a new one was
created. A new touch state was created if the old one is null. The
inverse of that can be used to detect if "switchedDevice" should be true.

Bug: 211379801
Test: m inputflinger_tests && adb sync data && adb shell -t /data/nativetest64/inputflinger_tests/inputflinger_tests
Change-Id: I57fad4d6c28409b89fb5c6c13fc37b930e4454da