Make servicemanager.microdroid uninstallable
`atest servicemanager_test` builds targets in the module directory,
which installs modules defined in the directory as well. This causes
servicemanager.microdroid to be installed and `m droid` next time will
result in a bad state (having both servicemanager instances with .rc
files)

Making the microdroid variant as `installable: false` so that it's not
installed in the target output directory. It's okay for microdroid
because the midrodroid build rule bundles it anyway.

Bug: n/a
Test: MicrodroidHostTestCases
Change-Id: I682c7beefd697f388694d6435491854597db1ab2