TouchInputMapper: Use ui::Transform to calculate orientation angles
Rather than manually re-orienting the calculated angles for orientation
and tilt, we use the transform to compute the oriented values for these
non-planar axes. This approach is both less error-prone and less verbose.

In this CL, we also transform the coverage rect from raw to display
space using the computed transform.

Bug: 236798672
Test: atest inputflinger_tests
Change-Id: Ibfb6d2ab43e6fd7a63736ee7d9610c42be44affd