Erase touch state if empty
Before this CL, touch state would remain stored even if it had no
windows inside. After a recent refactor, this started causing events
that have different sources to trigger the "conflicting pointer actions"
block.

The TooltipTest relied on this behaviour.

To fix this, erase touch state whenever there aren't any windows in it.
In the future, we may have to revise this test to send the correct
stream of events, which would include HOVER_EXIT.

Bug: 258845980
Test: m inputflinger_tests && out/host/linux-x86/nativetest/inputflinger_tests/inputflinger_tests
Test: atest android.view.cts.TooltipTest
Change-Id: I1cbe800793ba8cde55b81a3e1d1d4b0fe25f8c77