Do not link tests to libinputreader
In a recent CL, we started linking tests to libinputreader. This was
done because InputManager.cpp requires the symbol "createInputReader".
This symbol is defined in libinputreader.

However, instead, we'd like to build the tests from sources, so that we
don't need to reflash the device in order to iterate on the
implementation faster.

One thing that remains unclear is how this build used to work
previously, since InputManager.cpp always needed 'createInputReader'
symbol. The best guess here is that the compiler analyzed the
dependencies from the binary executable, and realized that
InputManager.cpp is not being tested / used, and therefore stripped all
such dependencies. Maybe for host, the compiler could not do that.

In this CL, we move InputManager.cpp directly into libinputflinger as a
source. This will also help highlight that this file is currently
untested.

Bug: 249591924
Test: m checkinput
Change-Id: I3b6741ecf6357623a7e684332e7eddc95425f903