Allow SkiaVkRenderEngine to not be compiled
A device will use RenderEngine with either SkiaVk or SkiaGL, but never
both at the same time.  Given that the choice is controlled by a
build-time value in the device's "device.mk" file, to save space, we
should allow the RenderEngine back-ends to be optionally built.  For
now, we'll only do this for SkiaVkRenderEngine.

Test: Compile
Bug: 259323498
Change-Id: Id547feaee7e8aeb4e46759d3d7c9c5dbe18f5211