TouchInputMapper: Consume hovering pointers outside physical frame
Similar to how we consume pointers that touch outside the physical
frame, we consume the hovering pointers when all the hovering pointers
are outside the physical frame.

This means hovering over the unmapped area of an external drawing
tablet will no longer produce hover events at the edge of the display.

Bug: 236798672
Test: atest inputflinger_tests
Test: manual with Wacom Intuos S
Change-Id: I6a2deb90311a2c9a98789ccd8318297cb2346208