libbinder_ndk: doc common stability error
Ideally we can store in Stability.cpp if stability gets
downgraded without the binder getting used, like other
Binder.cpp 'wasParceled' checks, but we have only seen
this issue one time. Putting a comment for now to
indicate difference (this is not an issue in C++ backend
because Bn* inherits from IBinder, only in NDK backend).

Fixes: 261199113
Test: N/A
Change-Id: I7e98eee2f830ccbf58edb961a5c3fa156c623344