SF: Update RefreshRateSelectorTest to check range of refresh rates.
We have increased number of Refresh rates with the
frame rate override being 1Hz and this test checks the
total number of refresh rates possible in tests. And
Tests a range of refresh rates from the start and end
of the refresh rate range for validation. The whole range of
refresh rates are high and will only increase going forward.
So having a range check helps with reducing the hard coded
number of refresh rates.

Test: atest libsurfaceflinger_unittest && atest SurfaceFlinger_unittest
&& atest libcompositionengine_test && atest CtsGraphicsTestCases && go/wm-smoke
BUG: 261520501

Change-Id: Iff0f2e28d2a80cfc46ccf993f1cb33affaee8440