Remove RE_SKIAVK ifdefs (always Compile SkiaVkRenderEngine)
This reverts 2 previous commits:

- "Allow SkiaVkRenderEngine to not be compiled" (commit
  646ebf21377ff28f9604846351a1230f9cd5c1a9)
- "Compile SkiaVkRenderEngine so that it can be enabled" ( commit
  32f9e69927c2efefe6f7a138bfd291f130ccc486)

that allowed SkiaVkRenderEngine to be conditionally compiled and then
set the define so that it would always be compiled.  We have decided
for now to have both SkiaGL and SkiaVk in RenderEngine, and so
removing the ifdef's makes the code simpler.

Note: the total increase in pinned memory that was observed in testing was between 360-362 KB (see b/263443538)

Test: `git revert` and visual inspection with gitk
Bug: 236390072
Change-Id: Ia49a2971c544db1d6b9a8996220e2b3e2d0ab437