Provide prediction time from the application
After further discussion with the graphics team, it turns out that the
Choreographer isn't the best way to compute the prediction time.

There may be some strategy that would use the Choreographer, like
looking at the last vsync and trying to guess when the next drawing will
occur.

However, for now, let's just provide this time from the application and
experiment with it. If we do need the Choreographer, we should be able
to get this data from libgui here in the native layer, so we don't need
the java to set this time, anyways.

Bug: 167946763
Test: atest libinput_tests inputflinger_tests
Change-Id: I496a9c184745d4cec085b880f913983cc2be9e3f