Bug reporting flow to work for shell running on any user
In the existing implementation dumpstate prompts a consent for all
the bug reports triggered from sources whouse calling_uid is not
AID_SHELL(2000), but with the recent refactorings for headles shell
can be running on any user and have different calling UID than 2000.

Change all the calling_uid assertion of shell to app_id to match with
shell process running on any user.

Refactored sendBroadcast method to send broadcast to the calling user.

Bug: 260989962
Test: manually tested consents are not triggered for the bugreport
triggered from shell on a secondary user.
Ignore-AOSP-First: will do it later with other headless changes.

Change-Id: I37dd1c5fbe68b3888ecec19533e19728c3f7d858