clang-format: PointerAlignment LEFT
This is the format used in libbinder, but it appears not to be
configured here. However - this is really strange - because
I know it must have been? Maybe the Google format changed?
However, I find it really shocking that we might have always
typed this correctly in libbinder. Certainly something
else broke the format here!?

Bugs: me
Test: N/A
Change-Id: I3f54bf359378b2080f9cc6e375d69c7e050b7d2d