Disable new touchpad stack for Sony gamepads
The touchpads on the Sony DualShock 4 and DualSense gamepads do not work
with the gestures library, for a reason that is currently unknown. This
is tracked in b/246587538, but while we wait for that to be fixed just
check for them by vendor and product ID, and fall back to the old stack.

This shouldn't really be a few IDs hard-coded into `InputDevice` — it
should be an input device configuration property, but since those are
currently only loaded in the configuration stage that will require a
larger change.

This is a second attempt after ag/21019000 caused a hwasan failure,
which we believe was due to calling getDeviceInfo() when not all mappers
had been configured. Getting the device ID from contextPtr instead fixes
this issue.

Bug: 251196347, 246587538
Test: `setprop persist.input.touchpad.gestures_library.enabled true`,
      connect gamepads, do a pinch gesture, confirm that touch spots
      show on screen (which doesn't happen in the new stack)
Test: `setprop persist.input.touchpad.gestures_library.enabled true`,
      run android.hardware.input.cts.tests.SonyDualshock4BluetoothTest
Change-Id: Ie843a7552ae31db640a847a9a108a66f6363730e