BBQ: Check if we have already acquired max num of buffers
When the IGBP is disconnected, the BQ state is reset but the
BBQ state is not. Without the max acquire check in BBQ, we
will continue to acquire buffers. This may cause issues in some
stress tests which continuously disconnect and reconnect while
queuing buffers. The buffers will only be released once they are
released by SF. This can cause some systems to run out of memory.
Ideal fix would be to track and free the buffers once the BQ is
disconnected but that will be more risky for QPR.

Test: atest GLSurfaceViewTest#testPauseResumeWithoutDelay
Bug: 263340543
Change-Id: I77b34f6151a9d1b461649c575be98df1f00f2464
Merged-In: I77b34f6151a9d1b461649c575be98df1f00f2464