Clear TrustedPresentationListener when a new listener is set
If a TrustedPresentationListener is overwritten for a Layer, the last
trusted presentation states need to be cleared.

It also needs to make sure composite is called after setting a
TrustedPresentationListener since the layer may already be in the
presented state and there's nothing new to compose. We need to make sure
the presentation state is computed at least once after setting the
TrustedPresentationListener

Test: TrustedPresentationCallbackTest
Bug: 256993331
Change-Id: I5cd344e9609d96c961fc70093ea705ee48f46817