Scheduler: Prevent removing the last display
unregisterDisplay calls promoteLeaderDisplay, which assumes that there
is at least one RefreshRateSelector remaining. Other Scheduler code also
assumes that there will be at least one display, which is necessary for
there to be a leader.

In SchedulerTest, remove the final call to unregisterDisplay, so we
don't put Scheduler in a bad state. Remove hasRefreshRateSelectors(),
since this condition must be true.

Bug: 241285191
Test: libsurfaceflinger_unittest
Change-Id: I827f20acd6a105d9f99e1ca9a7bfc59e633ab33f