SF: Pass latch time for bufferless surface frames
Latch time is used to classify jank type as BufferStuffing.
This jank classification does not  count as missed frames.
Layers without buffers do not pass in a latch time to the
frametimeline logic. Fix this inconsistency so we do not
incorrectly report missed frames for missed layer updates
due to buffer stuffing.

Test: check perfetto traces and see leashes are also classified as
buffer stuffing
Fixes: 266666415

Change-Id: Ie211aa3bd5821f6052cf84a62a2e245132a19d90