Ensure real touch cancels injected events
Sometimes, tests inject events globally. If a test is not hermetic (or,
it crashes), it may leave the system in an inconsistent state. To the
dispatcher, it would look like the virtual device is still down.
Currently, we don't subscribe to any injector death notifications (and
don't plan on doing so).

After this happens, the user may want to interact with the device, so
the notifyMotion call would come in. In this test, make sure that the
injected event is canceled when the new gesture is started.

Bug: 266382436
Test: m inputflinger_tests && $ANDROID_HOST_OUT/nativetest64/inputflinger_tests/inputflinger_tests
Change-Id: I9eeacd1e48cb1cd36a55acc8a129a83edd2784fc