[BBQ] Only update destination frame when buffer matches requested
The current code in BBQ won't reject the buffer if the buffer size and
last size match, even if requested size doesn't match. This is fine as
long as we don't also update the destination frame to the new requested
size. We can apply the old buffer, so we don't drop additional buffers,
but keep the last destination frame as the last requested size. When we
get a new buffer that matches the new requested size, we can update the
destination frame.

Without this fix, the buffer could be stretched in FREEZE mode since the
buffer size will end up not matching the destination frame set in SF.

Test: SurfaceSyncGroupContinuousTest
Fixes: 267284591
Change-Id: Id8bc8b6e04fa1cc56f7b9eb609cadb208b27b3ea