Allow debug logging of dispatcher's inbound events to be enabled on the fly
Enabling debug logs via adb commands required a reboot of the shell. In
order to debug flaky tests more easily, we allow raw event logs to be
enabled and disabled without the need for a restart on debuggable
builds.

We limit this ability to debuggable builds ensure there is no
performance overhead on user builds.

This change allows us to enable more verbose logging on flaky tests that
are hard to reproduce.

Bug: 193231132
Test: Manual, using adb command to enable debug logs on userdebug builds
Change-Id: Idf6f7fb35ad02e059582b1f0e7b3b2d006ff0730