Handle back shortcuts in native and remove home shortcut handling
Home: This is fully handled by System and can be handled in
PhoneWindowManager like all the other shortcuts.
Note: We are knowingly removing Meta+Enter for home which is the
old shortcut and is replaced by Meta+H.

Back: This event need to be sent to application for handling. Since,
meta based shortcuts or KeyEvents are not sent to application, we
need to pre-convert Meta+` and Meta+arrow_left to KEYCODE_BACK before
sending to App.

Test: atest ShortcutDispatchTest
Bug: 270242945
Bug: 270241579
Change-Id: I7dc17891920f7a5d61b4b3172533610a6dd4d65e