atrace.rc: fix duplicated block due to automerging
My aosp/2449540 was trying to sync aosp and internal state, but the
diffs lined up in a way that some internal branches got a duped
boottrace block. Resolve that manually with a merged-in spec with an
internal/master change-id as that branch and its downstream is
undamaged. Branches in question:
  https://screenshot.googleplex.com/82u82dZsAx9KjQG

Bug: 270060883
Merged-In: I6001eec487932ab1fd869231815fbf485b7370bb
Change-Id: I32a945f4d2cbb8337de7faf3c9e47fcc65bac089