RPC Binder: dropping all binders drops session
Previously, RPC binder was prone to leaks when you have
an incoming threadpool. Even if you dropped all references
to all binders, you still needed to manually call
shutdownAndWait.

Now, when you drop the last binder for a connection (after
you call getRootBinder the first time), it will automatically
shutdown the connection.

Previously, when this situation would happen, you could allow
a session to call getRootBinder again. However, we don't expect
or need this, so we can change the behavior.

Bug: 271830568
Test: binderRpcTest (+= 5s)
Change-Id: I2033f0f2908e238c0b96060489f72780b51a4bbe