Remove some old TODOs from GestureConverter.cpp
The first (about supporting more gesture types) is obsolete because we
now support all the gesture types we need. The other two sadly won't be
fixed, since the Gesture library is rather inconsistent about the signs
on its dx and dy values, and if I recall correctly the signs can't be
changed using gesture properties for multi-finger swipes.

Bug: 251196347
Test: none
Change-Id: Ib1e2e41a967a6675b258f415c1b411d37da853ec