RPC Binder: Retry Trusty connections several times
Sometimes the restarting of the Trusty service races
with tipc_connect, causing the latter to return -ENOTCONN
since the service has not come up yet. This puts the
connection attempts in a loop and retry them up to 5 times,
sleeping for a short duration between attempts.
Should make binderRpcToTrustyTest a lot less flaky.

Bug: 230135749
Test: run binderRpcToTrustyTest several times
Change-Id: I2c4e8dfe929535dbb77b5af3c677e5d27f00f50e