Generate texture pool asynchronously
This improves performance by moving allocateHelper to a different
thread, which is launched early so that the allocation is ready when
it is needed. In addition, each display will only need at most one
additional texture per frame, so replace the idea of a minimum pool
with making sure there is either at least one texture in the pool
or a future which will generate one.

Bug: 256184546
Test: atest libcompositionengine_test
Change-Id: Icb92a85c07b9f2911d1365aa9bbc8d0750a8c116