Refactor HardwareStateConverterTest
In a follow-up CL, I'll be modifying HardwareStateConverter to need more
constructor parameters, so moving the initialization into the test
constructor will make the tests a bit cleaner.

This removes the ability to add the MSC_TIMESTAMP axis before
initializing the converter in the MscTimestamp test. However, none of
the code under test actually depends on this axis being present in the
EventHub.

Bug: 259547750
Test: atest inputflinger_tests:HardwareStateConverterTest
Change-Id: Id99218c4dc81ef80014512489306465aee9fe427