Skip renderCachedSets if the Output is not enabled
This method was being called for a display that had not turned on yet.
The call to getSkipColorTransform resulted in a log warning - the call
only checks a cached value, and the value is only cached when the
display is first turned on. Skip the call when it is not enabled (e.g.
when it is off), removing the log spam.

Bug: 277358841
Test: look at logs
Change-Id: Ic5eb3a69d50719cf68e3b41d6eca659bd712f65f