Trigger system key user activity during keyguard
This CL changes the user activity handling in input so system keys will
always trigger user activity bypassing the DISABLE_USER_ACTIVITY flag that is used in the keyguard.

With the current information at our disposal, we believe that the flag
in question was added to avoid accidental screen touches from waking
up the device and consume the battery. However, this concern does not
exist with system keys as accidental presses should be a lot rarer.

Bug: 272052147
Test: Flashed and tried using system shortcuts
Change-Id: I89c3402a6ae4c2c75927e548ff38c871eb60479c