libbinder: remove 32-bit ABI support
This was an old userspace ABI, but we've encouraged all
platforms to move to a consistent ABI, regardless of
bitness of the device. This corresponds to a kernel
config option that we started requiring not be set
in Android P.

After this, the build system still needs to have this support
removed.

Bug: 232423610
Test: boot cf
Change-Id: Iff5997f910d86c73bbd6502ec393434d248cdcd5