Before deleting callbacks, ensure they aren't running
I3c1eccb36914f29560600d48bb08b1b8f2fe7c96 makes sure that a
VSyncDispatchTimerQueue unregisters its callback if a new one is moved
into it, but it's possible for the class to be used directly, so there
is still the possibility of deleting the callbacks while running.

Update the destructor to wait for any callbacks to finish before
deleting them. Also log an error, since the caller should've
unregistered them.

Bug: 279209321
Test: infeasible
Change-Id: Ia2c5436944c88284912be4963740be41de5c21d3