Add sanitizers to libinput
The library was already getting sanitized for some undefined behaviour on device. This change
will move the sanitization to all platforms like host, and also expands the list of sanitizers.

We can't enable the address sanitizer today, unfortunately. It causes the libinput to not be
able to load on some configurations. It's not clear what's causing that yet.

Bug: 271455682
Change-Id: I9588b0f7b42eec2070cd5556f94671fd934b7bec
Test: presubmit