Ensure injected touches poke user activity
Sometimes, there is a need to simulate a touch event via an adb shell
command. This could be done with the following command, for example:

adb shell input tap 500 500

In this CL, we add a unit test to check that the user activity is
getting poked after an event is injected.

Bug: 281559590
Test: m inputflinger_tests && $ANDROID_HOST_OUT/nativetest64/inputflinger_tests/inputflinger_tests
Change-Id: I6ac049135dddd828870a926906f28ba4239d9400