Slip touch when the new window is valid
Previously, the condition for slipping touch was based on the difference
in window tokens. However, that check missed the possibility of the new
window handle being null.

In this CL, that check is returned.

To reproduce this crash, touch needs to try to slip from a slippery
window into a window that requests to drop input.

Bug: 281601168
Test: m inputflinger_tests && $ANDROID_HOST_OUT/nativetest64/inputflinger_tests/inputflinger_tests
Change-Id: I0add7ee65a6492695e9b2368c83fe553677a32f6