Set cursor position to the first pointer for mouse events
In tests, a common task is to create an instance of NotifyMotionArgs.
A builder is used to define it.

Before this CL, the builder used to set cursor position only for the
cases where there is exactly 1 pointer in the event.

However, now, with the new touchpad stack, there may be > 1 pointer
created for the SOURCE_MOUSE events.
Relax the restriction to avoid the boilderplate of setting the
cursorposition for mouse events in the tests.

This patch is a temporary workaround for the ubsan failure in a newly
added test TouchpadThreeFingerSwipeNotSentToSingleWindow.

Bug: 263319225
Test: m inputflinger_tests && $ANDROID_HOST_OUT/nativetest64/inputflinger_tests/inputflinger_tests
Change-Id: Ifa8858c3f0f22978de822473d2362e70aba4e01c