Revert "Revert "Do not allow wildcard matching in GL loader.""
This reverts commit 237ab0873378805fc2d00fda9b3969216aad912e.

Reason for revert: The original patch is not the culprit.

Original commit message:
```
Original commit message:
```
Previously when ro.hardware.egl and ro.board.platform were not set, the
loader would attempt to load the exact GLES drivers, and if it failed,
it would attempt to use wildcard matching to find the GLES drivers.

However, ro.hardware.egl must be set to point to the GLES drivers and
hence wildcard matching should be disallowed. This patch makes sure the
GL loader no longer uses that path if a device is launched in Android 14
and beyond.
```

Reason for revert: startup regression in b/283393625
```

Bug: b/283393625
Change-Id: Id6fdd60e9a011679cedbd4eef5b3c7ea83882af9
Test: N/A