Fix the missing extension EGL_ANDROID_image_native_buffer
In http://ag/22040002, EGL_ANDROID_get_frame_timestamps
is conditionally added to the mExtensionString. A space
at the end of the extension name is mandatory, because
we use ' ' or '\0' as an indicator when parsing the
mExtensionString to find each individual extension name.

Bug: b/269060366
Bug: b/280016556
Test: m; atest android.hardware.nativehardware.cts.AHardwareBufferNativeTests
Change-Id: I2552b47dda49a09218a7741c45641c02bdd807ee
Merged-In: I2552b47dda49a09218a7741c45641c02bdd807ee