Improve input event tracing
Added the seq number to the server dispatch and client
received/finished trace events. This will enable correlating server
an and client events for easier input latency tracking/ANR debugging.

As a follow up, I'll consolidate some of the input trace events to
reduce pressure on the ftrace buffer. For now, this change should be
simple enough and low-risk to go into Android U.

Test: Manual
Bug: 285553040

Change-Id: Ic50235cfe854357f0c13ce62b904077355c780b9