Link libinput as a shared library for host
Since the rust dependencies in libinput are included as a
whole_static_lib, if we add the same rust dependencies in another
library (e.g. in libinputflinger), they will be included more than
once in the binary, resulting in duplicate symbol errors from the
linker. We can add it as a shared lib dependency to get around
the duplicate symbols.

Bug: 278783893
Test: build
Change-Id: I92897f3465852d7be030dbce995435edff9b54fd