Fix the bug in image compression that failed to specify correct usage, type, flags
The previous implementation of the vulkan swapchain incorrectly implemented
vkGetPhysicalDeviceSurfaceFormats2KHR. As a workaround at least one
driver vendor resorted to a code workaround in their driver to correct
the bad call in flight.

The failure was during a call to vkGetDeviceImageFormatProperties2 where
the swapchain was failing to include the correct image usage and image
type, along with a failure to include the correct image compression
flags.

This CL amends the mistake to add these missing components to the vulkan
vkGetDeviceImageFormatProperties2 call.

Any driver vendors that have not yet finalized their image compression
implementations will benefit from not needing to add an android specific
workaround for this bug. The CL as written does not require the driver
workaround that was already written to be revised. The driver will work
correctly without source changes. When the driver is updated the
workaround can cleanly be removed without side-effects.

Bug: 288100893
Test: dEQP-VK.api.image_compression_control.swapchain.android
Change-Id: Id10e5f3e42ae4e16b2cf4874a0ff4e0e2f21cd2d