Increase timeouts for ANR tests
This should help reduce the flake rate.
The flakes are not reproducible on host locally, but are flaking
remotely, with an overall flake rate of ~ 1% or less.

In the future, we can consider refactoring the dispatcher by providing
it a clock source that we can control, in order to make these faster and
completely remove flakes.

Bug: 282837934
Test: atest inputflinger_tests
Change-Id: I12fdc3ee1a823ede0e90256c54833d8d52aed636