Add sys/cdefs and __REMOVED_IN define to looper.h
The looper.h header introduced __REMOVED_IN, which requires the
sys/cdefs.h header to provide definitions for the macro. Add them here
and update files that include looper.h to not provide their own
definitions of __REMOVED_IN. Further, update comments to clarify that
the additional defines are in place to satisfy builds that use glibc or
Windows/MacOS libcs.

Bug: 14257980
Change-Id: Ia524b7e180c5137e84c205a77dc219b9e9e41608