Add type alias for DeviceId
To facilitate the future transition to a strong type, add a "weak" type
for DeviceId in this CL.

This will allow us to fix some of the existing code before having to
refactor all of it in a single topic.

Also, in this CL, simplify the names of some variables in TouchState
because we no longer have a conflict with the 'deviceId' member.

Bug: 211379801
Test: none
Change-Id: I1f287ad62fa26ba4bbbbbf353bebfbf8ad38ff67