[sf] Update framerate correctly with new frontend
In order to optimize snapshot updates, we only
update the snapshot if the requested state has changed.
If the hierarchy changes, we force update the children's
snapshot but this is not sufficient because a parent's
framerate might be affect by the child's framerate.

To fix this explicitly update the framerate when the
hierarchy changes.

Test: presubmit
Test: atest android.graphics.cts.FrameRateOverrideTest
Bug: 238781169
Change-Id: Ia633e7800cf3169c944313b4220b889bec0c3c5f