EventHub: Ensure bit arrays are large enough to store all event codes
We were previously initializing bit arrays with a min size equaling to
the MAX value for the event type (i.e. initialize array of min size
KEY_MAX for EV_KEY). This is incorrect, because the max value could be a
possible valid value, in which case it could not be stored in the array.

The bit arrays should be initialzed with the min size equaling the count
of the values (CNT == MAX + 1; e.g. KEY_CNT).

Bug: 290938220
Test: None
Change-Id: I1fa748f521a8539da444a4ae6e31cc4fb49be138