Remove redundant check for blurRegions to force client composition
If there are blurRegions on a Layer, we *should* force client
composition, but this is already handled in
Output::updateCompositionState +
Output::findLayerRequestingBackgroundComposition. The check on Layer is
redundant, so remove it.

Moreover, the check on Layer doesn't do enough. It doesn't handle blur-
behind, and it doesn't force the Layers below to use client composition,
which is also necessary.

This allows us to remove blurs and client composition from protected
surfaces on devices that do not have EGL_EXT_protected_content. (See
I9763eb22884e611568b36b2e221ee6d75ec3363e.)

Bug: 196271643
Test: OutputUpdateAndWriteCompositionStateTest
Change-Id: I055d221b1fcb36fb5fbc6747c353d3fb72f4d684