[DeviceAware] Use AttributionSourceState in native AppOpsService checkOp, noteOp, startOp, finishOp methods
This allows us to reduce the number of parameters being passed around
and also enables easy addition of new parameters such as device id.

Note that this change is unflagged and is meant to be a pure refactor.
This is preparatory CL and no new methods are being added to AppOpsManager yet, but will be needed
to allow clients to pass in device id.

Bug: 299160174
Test: atest CtsAppOpsTestCases CtsAppOps2TestCases
Change-Id: Ic2fdb088b3a3913a510c8c73f24570147267efd0