Provide a consistent motion event stream in libinput_tests
When the InputVerifier is enabled, some of the input tests fail because
they provide an inconsistent motion event stream to the InputPublisher.

In this CL, those are fixed. This should help us enable event
verification in the future.

There are still some failures remaining due to zero pointer tests that
will be addressed separately.

Bug: 211379801
Bug: 271455682
Test: enable InputVerifier in the code
Test: TEST=libinput_tests; m $TEST && $ANDROID_HOST_OUT/nativetest64/$TEST/$TEST
Change-Id: If19ea9bc50d2191ba2bdce3aad35390889db7f95