SF: Fix mode setting for UDFPS with disabled AOD
When AOD is off, DM sends its policy to SF while the display is powered
off. SF defers applying the policy of powered off displays, assuming it
will be applied when the display becomes the active display upon power-
on, as a special case for foldables. However, in the case of UDFPS with
disabled AOD, there is no change in active display, so SF did not apply
the policy to switch to 120 Hz on the lockscreen.

Tighten the condition for skipping mode set to only affect foldables in
non-concurrent mode.

Fixes: 300189378
Test: 120 Hz for UDFPS when Smooth Display and AOD are disabled.
Test: 60/120 Hz still works on fold/unfold and concurrent mode.
Test: DisplayModeSwitchingTest.powerOffDuringModeSet
Change-Id: Iedd25f8caed69c1dd304ee9f6341dba6bcdc87d7