Mark all pointers in a resampled event as resampled.
Even if the coordinates for a pointer are not resampled, they will be
added to an event with a timestamp that doesn't match what the device
is reporting. Algorithms that care about the consistency of pointer
coordinates will want to handle these events in the same manner as
resampled coordinates -- otherwise it may appear as though the
pointer has suddenly stopped moving.

Bug: 301277887
Test: atest --host libinput_tests
Change-Id: Idc833e9844856172ff749a90ee584292536524dc