Make the boot config tests more hermetic
Use a test harness to try to reset boot configs to its original state on
completion. Otherwise we would be polluting the state of other tests.

Ideally these tests hit a fake HWC instead and leave hitting a real HWC
instance for xTS. Alas, we don't have a convenient fake set up, and the
alternative is deleting these tests which is worse.

Bug: 304555116
Test: SurfaceFlinger_test, followed by inputflinger_tests
Change-Id: Id70d11713aca69c9da9f404060f377b2b0d2629e