Store isTrustedOverlay before returning early for hidden layers.
When a layer is hidden, we will still gather input info, but not
continue down the updateSnapshot method. Extracting isTrustedOverlay was
happening after so if the caller was a hidden spy window that also was a
trusted overlay, it would crash InputDispatcher since it would receive
the hidden windowInfo data with spy but not trusted overlay.

This fix moves the isTrustedOverlay extraction before the hidden check
so input info will contain trusted overlay info even if the window was
hidden

Test: LayerSnapshotTest#setTrustedOverlayForNonVisibleInput
Fixes: 304391559
Change-Id: Ie068c179635998e189814448062d86db0b089f40