libbinder_ndk: doc waitForService waits
This doc incorrectly stated before that this API will wait forever
if a service is not declared, but actually it has nothing to do
with whether the service is declared. The idea was that if a service
is not declared (and therefore not installed) it will wait forever.
However, we should just not mention whether the service is declared
and focus on whether the service is installed or not because that
will determine whether this API returns.

Bug: N/A
Test: N/A

Change-Id: Id07b5b5653a5b40b4c0d5a7e4f88e6b2440fff0a