include system ANGLE usage as ANGLE usage
This CL set the system ANGLE usage as GpuStatsInfo::Driver::ANGLE.
Previously it was categorized as GpuStatsInfo::Driver::GL.
The reason is ANGLE will be shipped as a system driver, not as APK.  We
want to monitor the adoption of system ANGLE driver.

Test: collect the GPU stats before and after the CL.  Check the GPU
stats reflects ANGLE traffic.  See details in b/294867368#comment17

Bug: b/294867368
Change-Id: I5fcc6a9b86f186ec5058cd3b2ddc262b4ef1b352