Downgrade fatal checks to warning
These checks were previously ALOG_ASSERT, which is actually a non-fatal
assertion that does not cause a crash.

In ag/If2eae87bc2a40b61144ddcd019a9800c2d526072, the function
addWindowTargetLocked was forked into the pointer and non-pointer
versions:
1) addWindowTargetLocked (non-pointer)
2) addPointerWindowTargetLocked (for pointers specifically)

At the same time, these checks were converted into LOG_ALWAYS_FATAL.
This has helped during development of that CL, because it found several
paths where dispatch modes were getting altered unexpectedly.

Unfortunately, we are now receiving reports of crashes in these areas.
That means that our current test coverage is insufficient.

In this CL, the checks are getting downgraded to warnings. This will
help us focus on finding ways to reproducing the crashes that we
received so far. In the future, we can look into strengthening these
checks again.

Bug: 308302190
Test: TEST=inputflinger_tests; m $TEST && $ANDROID_HOST_OUT/nativetest64/$TEST/$TEST
Change-Id: I70f92d5576d733a1ba423115026c4ff66da8e682