include system ANGLE usage as ANGLE usage
This CL set the system ANGLE usage as GpuStatsInfo::Driver::ANGLE.
Previously it was categorized as GpuStatsInfo::Driver::GL.
The reason is ANGLE will be shipped as a system driver, not as APK.  We
want to monitor the adoption of system ANGLE driver.

Test: collect the GPU stats before and after the CL.  Check the GPU
stats reflects ANGLE traffic.  See details in http://b/308476674#comment3

Test: run CtsAngleIntegrationHostTestCases on ABTD with the change, See
details in http://b/308476674#comment2

Bug: b/308152854, b/308476674, b/308098783
Change-Id: I9beb7c75d65b7aa2b7c5af9f262f30be9d6247d5