libbinder_rs: Switch inner type of ParcelFileDescriptor
Change the inner type from File to OwnedFd to support
OSes that implement the latter but not the former.
An added benefit is that this brings the type closer
to the set of supported types that can be wrapped in
a ParcelFileDescriptors, e.g., sockets and memory buffers
in addition to files.

Bug: 242243245
Test: atest aidl_integration_test
Change-Id: I9ee14e479ab65c31459e6cecd736039deac1f29b