Only use InProcess backend in fuzzer
This is a partial revert of 7e6446c3da42("Refactored
surfaceflinger_frametracer_fuzzer").

I think the intent of the fuzzer is more to stress the surface flinger
tracing code and less to stress the perfetto consumer code (which is
less security critical).

kUnspecifiedBackend is equivalent to kInProcessBackend (it's the first
registered backend which we initialized earlier with
perfetto::Tracing::Initialize()). kCustomBackend will not have any
effect, because no custom backend has been previously registered.
kSystemBackend will make the producer talk with the real perfetto
tracing service on the system: this doesn't have much value because the
real code is not running under the fuzzer there. The only code under the
fuzzer will be the consumer part of the perfetto C++ SDK, which is not
security critical.

This makes the fuzzer run much faster.

Bug: 307601836
Change-Id: Ia5c4b42bf4916783d7eafa7b51b97ee6a7fc726e