Update lastExpectedPresentTime when cadence changes
When cadence changes, updating the
lastExpectedPresentTime prevents from
having an expected present time from a different cadence.

From getDeviceCompositionChanges we only store
the future expectedPresentTime.
getDeviceCompositionChanges follows after
onVsyncGenerated and stores the
expectedPresentTime otherwise from the past when rate changes.
This causes the next onVsyncGenerated to think
that the rate has changed, while it stayed the
same, and SF is flushing rest of the buffers
at the previous rate.

BUG: 296636253
BUG: 296635769
Test: manual
Change-Id: Ifb45d875e381a737b5e944d2e6cb080c5eb81196