Move hover rejection to an earlier part of the function
Hover events from the same device are rejected if this device is already
down.

Before this CL, this rejection took place later in the pipeline, after
many other computations were already performed.

In this CL, this check is going to be done earlier.

This helps separate the logic a bit, because now there will not be any
return statements after the injection result is set to succeeded.

Bug: 211379801
Test: TEST=inputflinger_tests; m $TEST && $ANDROID_HOST_OUT/nativetest64/$TEST/$TEST
Change-Id: Ib0f1ad750a8bd172decc63c82289ef014b8e470d