SF FlagManager: ignore debug override during tests
Tests use SET_FLAG_FOR_TEST, which specifies the value of the flag that
the test is expecting. But if the debug override is set, the test will
test under the wrong conditions.

In addition, there is no need to modify the static value in unit test
mode. Simply return the value returned by the flag system.

Bug: 241285491
Test: libsurfaceflinger_unittest:FlagManagerTest
Test: libcompositionengine_test after
      Ie2836e07e900c0269d60112ec0ebfa82ee4b82d7 with different values
      set for debug.sf.multithreaded_present
Change-Id: Ide557e06ed7511b47033c2cb7016ad81bac40dad