Immediately fail screenshots if requesting secure without permission.
The screenshot logic will only fail to capture when requesting secure if
there is any secure layers on screen. This is confusing since the caller
shouldn't be allowed to apss in captureSecureLayers for screenshots
without permission regardless if there are any secure layers on screen.

This makes the screenshot permission model simpler
1. If there are secure layers on screen, it will be blacked out
2. If the caller has CAPTURE_BLACKOUT_CONTENT and requests
   captureSecureLayers, they will get a screenshot that contains the
   secure layers
3. If the caller doesn't have CAPTURE_BLACKOUT_CONTENT and requests
   captureSecureLayers, they will get a permission denied callback.

Test: ScreenCaptureTest
Bug: 313697941
Change-Id: Iba64989d74ebc2076218f643a8940806e59b6b97