rust: Use wrapper for header file
With the update to C++20, we see warnings for #pragma in a non-header
file. The entry file for a bindgen compilation is not treated as a
header file, which raised this issue.

This approach has the added benefit that rather than using an explicit
filegroup, it uses the headers exported by the library they're binding
against.

Bug: 311052584
Test: mma in binder folder
Change-Id: Ib0456db5a09f7e5225001837977417186db33876