Pointer Icon Refactor: Show cursor immediately when mouse is connected
When a mouse or touchpad is connected, show the mouse cursor
immediately, instead of waiting for the first event to be generated by
the device.

This also means that the mouse cursor position will be valid whenever
there is a mouse or touchpad connected, whereas before, the cursor
position would be invalid until the first event is generated by the
device.

Bug: 315509086
Bug: 290924778
Bug: 293587049
Test: atest inputflinger_tests
Change-Id: I55898a3de1beb0f83f5da199521f26a886fb596c