CursorInputMapper: move unit tests to InputMapperUnitTest
This puts most of CursorInputMapper's unit tests in one place
(CursorInputMapper_test.cpp) and under one input mapper testing system,
which will hopefully simplify future additions to the tests. The only
CursorInputMapper tests remaining to be migrated are those which rely on
setting an associated viewport, which I'm saving for a follow-up CL
since they'll need the introduction of a fake InputDeviceContext.

Also make a few smaller clean-ups:

* underscores have been removed from test names, as they can cause
  problems in gTest;
* assertions have been moved to the newer style using matchers,
  improving conciseness and readability;
* tests for the creation of NotifyKeyArgs for certain mouse buttons have
  been parameterized.

Bug: 283812079
Test: atest inputflinger_tests
Change-Id: I0f0307dc57edcc62d0ce24b15bf0d9e8b8b915c6