Avoid capturing always on tracing when generating layers trace
When we encounter unexpected states, we write the always on
transaction traces to disk so we can investigate and fix these
issues in the field. The states also occur when we generate layers
trace since we use the same logic to replay back the transactions.
In this case we want to avoid writing the trace to disk. This is only
an issue with the perfetto backend since it converts the traces
in process.

Test: presubmit
Fixes: 316896588
Change-Id: I8729328c556f5c0575ac8879d51529cc51259bac