bufferstreams: Add AIDL interfaces/parcelables for stream types
These interfaces incorporate jreck@'s ideas for caching and
BufferAttachements. This means that we only need to send the
heavy-weight HardwareBuffers once, and then we can just send a
lightweight reference to the buffer.

This interface is not intended for public use yet and is going to be
actively developed for the next few months.

Bug: 296099728
Bug: 296450069
Test: builds
Change-Id: I6ca17dda1186ae74773bc344f3cda4b5574585aa