Deflake ExternalStylusIntegrationTest test case
The integration test was testing a case where the code is expected to do
nothing until a 72ms timeout expires, and then checking that an event is
generated.

To verify this, the test would wait exactly 72ms, and assert there were
no events generated. Since this is an integration test, waiting the
exact amount of time as the timeout used in the code is naturally flaky.

Resolve this by using a smaller timeout that the one used in the code
when enforcing that an event is not generated.

Bug: 304787002
Test: atest ... --rerun-until-failure
Change-Id: Ib0bcc6de149efee575d9fda6482fffaa2edfa15d