Add slow_keys input filter
DD: go/pk_accessibility
‘Slow keys’ is an accessibility feature to aid users who have
physical disabilities, that allows the user to specify the
duration for which one must press-and-hold a key before the
system accepts the keypress.
Note: As part of this CL, introduced threading in input_filter
stage in RUST. Due to Jni requirements, have used the utils/Thread.h
impl to create a thread and pass it over to rust for handling.

Test: atest --host libinputflinger_rs_test
Bug: 294546335
Change-Id: Ib418c133e09f3085e5fe78991068d4be9ed09ed9