Cleanup Scheduler::initVsync
Oddly, we pass a member of Scheduler to a method on Scheduler. This
method is actually on Scheduler's base class, MessageQueue, which does
not have that member. Rename the method to initVsyncInternal, and add
a new method, Scheduler::initVsync, which passes its member to
MessageQueue::initVsyncInternal. ("Internal" is needed to prevent
shadowing a virtual method.)

No change in logical behavior.

Bug: 266816604
Test: mmma -j frameworks/native/services/surfaceflinger
Change-Id: Ia7fcb9d76638bbce3051d4778defd10bd5c86ad1