Add captureLayersSync function
This captureLayers function in SurfaceFlinger will wait on the
requested binder thread for the screenshot composition to complete
before returning the buffer to the client. This is different than the
existing captureLayers method since the other request is oneway and
invokes the screen capture callback on a different binder thread.

This is needed because there are places in system server that request
screenshots while holding a lock and then wait synchronously on the
results. While waiting on the buffer and holding the lock, additional
two way binder calls can be made into system server that are waiting to
acquire the same lock. If there are enough requests, we may run out of
binder threads and then the screen capture result can't be posted back
to system server because it needs a free binder thread. This will result
in a deadlock because the lock that the screenshot request is holding
can never be unlocked without a free binder thread. Binder threads will
never be freed up because they are waiting to acquire the lock.

The async screencapture code is still useful for cases where there's no
global lock being held while waiting on results or the results is posted
to another thread.

Test: Screenshots
Bug: 321263247
Change-Id: I259173a59f488995e13af8f7dd2ca98c3bbf8639