Add another pragma to disable another bunch of warnings.
I suspect someone intended there to be an AIOOBE check in here, but I
can't understand the code well enough to be sure of that, let alone well
enough to actually implement it. So just add this warning to the list of
warnings we're already ignoring.

Also update the generator to take into account the stuff that folks have
manually added to the generated files... There's still one manual edit
necessary because the generator can't cope with the fact that
eglDupNativeFenceFDANDROID() has a hand-written Java wrapper, so the
generated native method needs to be private and renamed. But at least
we're down to one trivial manual edit (adding a line to the methods[] in
android_opengl_EGLExt.cpp).

Test: treehugger
Change-Id: I9f80315612192d65f1e2710a53a9404757ae34bd