Revert "Skip a step when blurring"
Modified revert of I4caa7ef3b5357c723fdc81c8e18b159c87366b60.

The thinking behind that CL was that we could skip drawing the non-
blurred image, because we would draw over it anyway with the blurred
image.

But during rotation and other animations, we don't always cover up
the old drawing, leading to old images showing up. This is particularly
jarring on a tablet.

Put the change behind a flag.

Bug: 236231351
Fixes: 255921628
Test: manual
Test: librenderengine_test
Change-Id: Ie9969b17b49964395f2d055bbb610c516e5df10c