Increase the test timeout used when waiting for events to be traced
We need to use a larger timeout while waiting for events to be traced,
compared to the timeout used while waiting to receive events through the
input channel. Since events are traced from a separate thread that
doesn't have as high of a thread priority as the InputDispatcher's
thread, the tracer is expected to lag behind the dispatcher.

Bug: 323347575
Bug: 210460522
Test: inputflinger_tests --gtest_filter="*InputDisp*" --gtest_repeat=100 --gtest_break_on_failure
Change-Id: I7dfe648c4de5ef39f7c44ad81b544634a54e0e68