Remove TransferDeathRecipients when ABpBinder is deleted
Store a weak reference on the ABpBinder to the death recipients it's
linked to so it can initiate a clean up when being destructed.

Test what happens when a binder is deleted.
Test what happens when a binder is deleted from the death recipient when
the binder dies.

Ignore-AOSP-First: b/319210610

Test: atest libbinder_ndk_unit_test
Bug: 325612851
Change-Id: Ifc532b64d8ca0fb586f7fbadd0956e103ded94fc