InputReader_test: Wait for stylus gesture to be notified
This is required because of the following change:
I818b1e688565d30867295019c5baae29cb5bdbf2

Previously, the policy was notified of the stylus gesture changing
before the events were flushed to the listener. This meant the policy
was always notified of the stylus gesture by the time the listener
received the stylus DOWN event.

After the aforementioned change, this is no longer true, so we cannot
use the DOWN event as a sync signal. Instead, we need to wait for the
policy to be notified.

Bug: 324318955
Test: atest inputflinger_tests
Change-Id: I8219e0db092c14da5907337b3ea5b5c6f7dddd4a