Don't repeat when slow keys A11y feature enabled
Currently when slow keys is enabled, user has to press and hold
for the key to be registered, but since, we still haven't added
settings to modify repeat thresholds, usually it ends up repeating
the pressed key as well which may be annoying for the user.
As a stop gap, we block repeat for now when slow keys is enabled.

Future plan at V+ is to have repat threshold modification in
Settings allowing user to properly configure it as per the need.

Bug: 322327461
Test: Manual
Change-Id: I44503746c16ed992bbfbfe852e26e22896f32595