[SF] Change FrameRateCategory Max range to 120
When we cap the refresh rate we don't select the
higher refresh rate for the cases when multiple
layers vote for the lower rate.
This affects the frames which request the higher
rate along side the frames that vote for
lower rate.
This change will give the higher refresh
rate an equal amount of weight in this case.

Test: atest RefreshRateSelectorTest
Bug: 328084623
Change-Id: I8a48e2b066e8908ff8f7579615260d8c74a79bff