Rely on C++ wake and sleep calls for InputFilter thread
Earlier, we used rust thread park()/unpark() to put the thread
to sleep and wake up from sleep. But that caused some breakages
after migrating the rust system crates to 2021 edition.
Since, the threads are created in C++, it was more reliable to
rely on C++ side of the implementation to implement the sleep
and wake functions.

Test: atest --host libinputflinger_rs_test
Test: manual
Bug: 294546335
Change-Id: I2eecae4716529d9d4cdd70ab22e784afc68648bf