Spy may receive events out of order
In the current implementation, the spy window may receive events out of
order.

In the test being fixed, there are 2 input devices active. The spy then
pilfers the streams.
Afterwards, two additional MOVE events, one for each device, are sent.

We are observing that the spy may receive these two move events out of
order.
That's unexpected, because the order of events inside dispatcher and
inside the socket transfer should be well-defined.

However, that seems harmless, so allow this to happen while
investigation is ongoing.

Bug: 332314982
Test: TEST=inputflinger_tests; m $TEST && adb sync && adb shell -t /data/nativetest64/$TEST/$TEST --gtest_filter="*MultiDevicePilfer" --gtest_repeat=10000 --gtest_break_on_failure
Change-Id: I57f40965bb8297b62033b4bf0ea9326ecebfcfa0