InputTracingTest: Fix test flakiness
There was a bug in the logic for tracking when the tracing thread is
idle. The thread can only be idle when the queue is empty, so prevent
it from being marked as idle when there are still events left to be
processed.

Also, a waitForTracerIdle() was missing in the TraceWindowDispatch test.

Bug: 333544493
Test: InputTracingTest passes 10000 times
Change-Id: I90feaeefc80d1a0e43732ee7e75d4074a9634bd7