Add enabled state to InputDevice and remove isInputDeviceEnabled(2/n)
Since we have added a field for the enabled state to InputDeviceInfo, we
have also added the same status field to InputDevice.java. This way, we
won't need to call the isInputDeviceEnabled method in InputManager again
to send a request to InputManagerService to check if the device is
enabled. At this point, we can directly remove the isInputDeviceEnabled
method in InputManager, which is used as a hidden API, since it is no
longer in use.

Bug: 336420877
Test: build & atest InputTests

Change-Id: I7c2a600a55941dcb55c189c1e1f24e6787173d64
Signed-off-by: Linnan Li <lilinnan@xiaomi.corp-partner.google.com>