Fix VibratorCallbackScheduler destructor lock
The VibratorCallbackScheduler destructor joins on the scheduler
thread to wait for the main loop to finish, but the conditional variable
is waiting indefinitely without a predicate, which can cause it
sometimes to miss the notify call from the destructor and get stuck.

Adding a predicate condition fixes the VibratorCallbackSchedulerTest
flakiness for the timeout "No test results." failures.

Bug: 293603710
Bug: 293623689
Test: atest --rerun-until-failure 1000 VibratorCallbackSchedulerTest
Change-Id: Id9501c10fe5209003d9b74b0f39f2bcf87de05c2